权限和更改密码优化

+++ client/src/components/tagManager.vue
@@ -1,5 +1,5 @@
 <template>
-  <div>
+  <div id="tagAction">
     <el-row>
       <strong>标签管理</strong>
     </el-row>
@@ -36,8 +36,9 @@
     name   : 'articleManager',
     data() {
       return {
-        tags  : [],
-        userId: JSON.parse(localStorage.getItem('currentUser')).id
+        tags       : [],
+        currentUser: JSON.parse(localStorage.getItem('currentUser')),
+        userId     : JSON.parse(localStorage.getItem('currentUser')).id
       }
     },
     mounted: function () {
@@ -55,11 +56,11 @@
           let tag = {
             title: value
           };
-          this.$http.post(`/api/tag/create`, tag).then((res) => {
+          this.$http.post(`/api/user/${this.userId}/tag/create`, tag).then((res) => {
             this.refreshData();
             this.$message({
               type   : 'success',
-              message: '添加成功'
+              message: value
             })
           });
         }).catch(() => {
@@ -70,47 +71,55 @@
         })
       },
       editTag(id, title) {
-        this.$prompt('请输入标题', '编辑', {
-          confirmButtonText: '确定',
-          cancelButtonText : '取消',
-          inputValue       : title
-        }).then(({value}) => {
-          let tag = {
-            title: value
-          };
-          this.$http.put(`/api/tag/${id}/update`, tag).then((res) => {
-            this.refreshData();
+        if (this.currentUser.admin) {
+          this.$prompt('请输入标题', '编辑', {
+            confirmButtonText: '确定',
+            cancelButtonText : '取消',
+            inputValue       : title
+          }).then(({value}) => {
+            let tag = {
+              title: value
+            };
+            this.$http.put(`/api/tag/${id}/update`, tag).then((res) => {
+              this.refreshData();
+              this.$message({
+                type   : 'success',
+                message: '更新成功'
+              })
+            });
+          }).catch(() => {
             this.$message({
-              type   : 'success',
-              message: '更新成功'
+              type   : 'info',
+              message: '取消输入'
             })
-          });
-        }).catch(() => {
-          this.$message({
-            type   : 'info',
-            message: '取消输入'
           })
-        })
+        } else {
+          this.$message.error('您没有操作权限！')
+        }
       },
       removeTag(id) {
-        this.$confirm('确定删除该标签，同时删除该标签下的文章？', '提示', {
-          confirmButtonText: '确定',
-          cancelButtonText : '取消',
-          type             : 'warning'
-        }).then(() => {
-          this.$http.delete(`/api/tag/${id}/destroy`).then((res) => {
-            this.refreshData();
+        if (this.currentUser.admin) {
+          this.$confirm('确定删除该标签，同时删除该标签下的文章？', '提示', {
+            confirmButtonText: '确定',
+            cancelButtonText : '取消',
+            type             : 'warning'
+          }).then(() => {
+            this.$http.delete(`/api/tag/${id}/destroy`).then((res) => {
+              this.refreshData();
+              this.$message({
+                type   : 'success',
+                message: '删除成功'
+              })
+            });
+          }).catch(() => {
             this.$message({
-              type   : 'success',
-              message: '删除成功'
+              type   : 'info',
+              message: '已取消删除'
             })
-          });
-        }).catch(() => {
-          this.$message({
-            type   : 'info',
-            message: '已取消删除'
           })
-        })
+        } else {
+          this.$message.error('您没有操作权限！')
+        }
       },
       goBack() {
         window.history.back();
@@ -118,25 +127,33 @@
     }
   }
 </script>
-<style scoped lang="stylus">
-  .tool {
-    position relative
-    .goBack {
-      position absolute
-      right 5px
-      top 0
+<style lang="stylus">
+  #tagAction {
+    .tool {
+      position relative
+      .goBack {
+        position absolute
+        right 5px
+        top 0
+      }
+    }
+
+    .el-button {
+      outline: none
+    }
+
+    .circle {
+      padding 6px 6px;
+    }
+
+    .el-row {
+      text-align: left;
+      margin-bottom: 10px;
+    }
+    tr {
+      th {
+        text-align center
+      }
     }
   }
-  .el-button {
-    outline: none
-  }
-  .circle {
-    padding 6px 6px;
-  }
-  .el-row {
-    text-align: left;
-    margin-bottom: 10px;
-  }
-
-
 </style>
